Этап 2 (выполнен)

+++ core/request-service/src/main/java/ru/practicum/request/repository/RequestRepository.java
@@ -21,7 +21,7 @@ public interface RequestRepository extends JpaRepository<ParticipationRequest, I
 
     List<ParticipationRequest> findAllByIdIn(List<Integer> ids);
 
-    boolean existsByEventIdAndRequesterIdAndStatus(Long eventId, Long requesterId, RequestStatus status);
+    boolean existsByEventIdAndRequesterIdAndStatus(Integer eventId, Integer requesterId, RequestStatus status);
 
     @Query("SELECT COUNT(r) FROM ParticipationRequest r " +
             "WHERE r.eventId = :eventId AND r.status = :status")
